requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
apscheduler==3.11.0
pytest==8.3.3
